requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==5.0.6
googleapis-common-protos==1.63.1
grpcio==1.64.1
idna==3.7
importlib_metadata==7.1.0
opentelemetry-api==1.26.0
opentelemetry-distro==0.47b0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-asyncio==0.47b0
opentelemetry-instrumentation-aws-lambda==0.47b0
opentelemetry-instrumentation-dbapi==0.47b0
opentelemetry-instrumentation-django==0.47b0
opentelemetry-instrumentation-grpc==0.47b0
opentelemetry-instrumentation-logging==0.47b0
opentelemetry-instrumentation-psycopg2==0.47b0
opentelemetry-instrumentation-requests==0.47b0
opentelemetry-instrumentation-sqlite3==0.47b0
opentelemetry-instrumentation-threading==0.47b0
opentelemetry-instrumentation-urllib==0.47b0
opentelemetry-instrumentation-urllib3==0.47b0
opentelemetry-instrumentation-wsgi==0.47b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-test-utils==0.47b0
opentelemetry-util-http==0.47b0
protobuf==4.25.3
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
wrapt==1.16.0
zipp==3.19.1
